Add files via upload
As requested, attached is an asset that tells the player when they are nolonger on the menu.  As you can hear, there is a sound when they are not on the menu.   I have sent it to you as an asset so just put it in your assets folder and into a scene.  I am unable to send it as a project file for security and size reasons, and the build as well was too big.  I recommend making a window mode because the menu button in the middle does not do anything yet.  There is also not yet an exit button so to exit simply click on the normal exit button at the top right corner of your screen.   Hopefully this is what you were looking for.  Feel free to let me know if you have any questions.